Fixed an integer wraparound bug.  Wasn't affecting anything until the way, way late game, probably farther than anyone has ever been.

+++ src/org/quark/ogame/OGameROI.java
@@ -203,7 +203,7 @@ public class OGameROI {
 			if (hours < 0) {
 				return Duration.ofDays(365000000);
 			}
-			return Duration.ofSeconds((int) (hours * 60 * 60));
+			return Duration.ofSeconds((long) (hours * 60 * 60));
 		}
 	}
 }
